+++ assets/src/utils/color.ts
@@ -2,7 +2,6 @@ import { normalizeColor } from 'grommet/utils'
 import { randomColor } from 'randomcolor'
 import { useMemo } from 'react'
 
-// Colors used in graphs.
 export const COLORS = [
   '#99DAFF',
   '#D596F4',
@@ -35,6 +34,20 @@ export const COLORS = [
   '#F5E093',
 ]
 
+export const alpha = (hex, alph) => `${hex}${Math.floor(alph * 255).toString(16).padStart(2, '0')}`
+
+const coerce = color => (color < 255 ? (color < 1 ? 0 : color) : 255)
+
+export function shadeColor(color, percent) {
+  const num = parseInt(color.replace('#', ''), 16)
+  const amt = Math.round(2.55 * percent)
+  const R = (num >> 16) + amt
+  const B = (num >> 8 & 0x00FF) + amt
+  const G = (num & 0x0000FF) + amt
+
+  return `#${(0x1000000 + coerce(R) * 0x10000 + coerce(B) * 0x100 + coerce(G)).toString(16).slice(1)}`
+}
+
 export function generateColor(i = -1) {
   return i >= 0 && i < COLORS.length ? COLORS[i] : randomColor()
 }
@@ -42,3 +55,4 @@ export function generateColor(i = -1) {
 export function useColorMap(theme, colors) {
   return useMemo(() => (colors || COLORS).map(c => normalizeColor(c, theme)), [theme, colors])
 }
+
